backend/pyproject.toml
[project]
name = "jpb-ai-session"
version = "0.1.0"
description = "Jnj Python Backend for AI Session Management"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.29.0",
]
